search and get all stories

+++ routes/stories.js
@@ -35,6 +35,16 @@ router.post('/create', extractStoryFiles, verifyStoryInput, async (req,res) => {
             res.status(500).json({msg: err.sqlMessage});
             console.log(err);
         } else {
+            let keyword_sql = 'INSERT INTO keyword_to_story SET ?';
+            for (keyword of req.body.keyLearningOutcomes) {
+                let keyword_to_story = {
+                    story_id: result.insertId,
+                    keyword
+                }
+                db.query(keyword_sql, keyword_to_story, (err, result) => {
+                    if (err) console.log(err);
+                });
+            }
             res.status(200).send({uuid: result.insertId});
         }
     });
@@ -155,6 +165,10 @@ router.get('/favourites', auth, async (req,res) => {
 router.post('/favourite', auth, [
     body('storyId', 'Please provide a valid storyId').not().isEmpty()
 ], async (req,res) => {
+        const requestErrors = validationResult(req);
+        if(!requestErrors.isEmpty()){
+            return res.status(400).json({msg: requestErrors.array()});
+        }
         let sql = `INSERT INTO user_to_story_favourite SET ?`;
         let user_to_story_favourite = {
             user_id: req.user,
@@ -179,6 +193,10 @@ router.post('/favourite', auth, [
 router.post('/unfavourite', auth, [
     body('storyId', 'Please provide a valid storyId').not().isEmpty()
 ], async (req,res) => {
+    const requestErrors = validationResult(req);
+    if(!requestErrors.isEmpty()){
+        return res.status(400).json({msg: requestErrors.array()});
+    }
     let sql = `DELETE FROM user_to_story_favourite where user_id = ${req.user} and story_id = ${req.body.storyId}`;
 
     db.query(sql, (err, result) => {
@@ -198,7 +216,67 @@ router.post('/unfavourite', auth, [
 router.post('/search', [
     body('keyLearningOutcomes', 'Please provide valid search conditions').isArray()
 ], async (req,res) => {
-    res.status(200).send(`Returns stories with corresponding key words ${req.body.keyLearningOutcomes}`) ;
+    const requestErrors = validationResult(req);
+    if(!requestErrors.isEmpty()){
+        return res.status(400).json({msg: requestErrors.array()});
+    }
+    const keywords = req.body.keyLearningOutcomes;
+    let sql = `select * from keyword_to_story INNER JOIN stories on keyword_to_story.story_id = stories.id where keyword in (${keywords.map(keyword => `'${keyword}'`)})`;
+    db.query(sql, async (err, result) => {
+        if (err) res.status(500).json({msg: err.sqlMessage});
+        const coverPhotos = await Promise.all(result.map(story => {
+            return getS3Object(story.cover_photo_path)
+        }));
+
+        searchStories = []
+        const v = {};
+        for (let i = 0; i < coverPhotos.length; i++) {
+            const id = result[i].id
+            if (id in v) continue;
+            v[id] = true;
+
+            const story = {
+                title: result[i].title,
+                author: result[i].author,
+                description: result[i].description,
+                keyLearningOutcomes: JSON.parse(result[i].key_learning_outcomes),
+                coverPhoto: coverPhotos[i],
+            }
+            searchStories.push(story);
+        }
+        res.status(200).json(searchStories);
+    });
+});
+
+// @route POST /stories/search
+// @description Returns stories with matching key learning outcomes
+// @access Public
+router.get('/all', async (req,res) => {
+    let sql = `select * from stories`;
+    db.query(sql, async (err, result) => {
+        if (err) res.status(500).json({msg: err.sqlMessage});
+        const coverPhotos = await Promise.all(result.map(story => {
+            return getS3Object(story.cover_photo_path)
+        }));
+
+        stories = []
+        const v = {};
+        for (let i = 0; i < coverPhotos.length; i++) {
+            const id = result[i].id
+            if (id in v) continue;
+            v[id] = true;
+
+            const story = {
+                title: result[i].title,
+                author: result[i].author,
+                description: result[i].description,
+                keyLearningOutcomes: JSON.parse(result[i].key_learning_outcomes),
+                coverPhoto: coverPhotos[i],
+            }
+            stories.push(story);
+        }
+        res.status(200).json(stories);
+    });
 });
